+++ site-cookbooks/algieba/recipes/deploy.rb
@@ -60,12 +60,25 @@ deploy node[:algieba][:deploy_dir] do
       environment 'RAILS_ENV' => node.chef_environment, 'PATH' => node[:rvm][:path]
       only_if { node.chef_environment == 'production' }
     end
+
+    execute 'sed -i -e "s/<%= ENV\[\"SECRET_KEY_BASE\"\] %>/`rvm 2.2.0 do bundle exec rake secret`/g" config/secrets.yml' do
+      cwd release_path
+      environment 'RAILS_ENV' => node.chef_environment, 'PATH' => node[:rvm][:path]
+      only_if { node.chef_environment == 'production' }
+    end
   end
 
-  if File.exists?("#{node[:algieba][:deploy_dir]}/shared/tmp/pids/unicorn.pid")
-    restart_command 'rvm 2.2.0 do bundle exec rake unicorn:restart'
-  else
-    restart_command "rvm 2.2.0 do bundle exec rake unicorn:start RAILS_ENV=#{node.chef_environment}"
+  restart_command do
+    execute 'rvm 2.2.0 do bundle exec rake unicorn:stop' do
+      cwd release_path
+      environment 'RAILS_ENV' => node.chef_environment, 'PATH' => node[:rvm][:path]
+      only_if { File.exists?("#{node[:algieba][:deploy_dir]}/shared/tmp/pids/unicorn.pid") }
+    end
+
+    execute 'rvm 2.2.0 do bundle exec rake unicorn:start' do
+      cwd release_path
+      environment 'RAILS_ENV' => node.chef_environment, 'PATH' => node[:rvm][:path]
+    end
   end
 end