+++ site-cookbooks/algieba/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of the algieba cookbook.
 
+## 0.2.0
+- Fix bug to restart app
+- Create secret key base for production
+
 ## 0.1.2
 - Move mysql users to attribute and environments